API/app/core/jwt_auth.py
"""JWT creation and verification for student auth."""
from datetime import datetime, timezone, timedelta
from uuid import UUID

import jwt

from app.core.settings import settings


def create_token(learner_id: UUID, username: str) -> str:
    now = datetime.now(timezone.utc)
    expire = now + timedelta(minutes=settings.jwt_expire_minutes)
    payload = {
        "sub": str(learner_id),
        "username": username,
        "iat": now,
        "exp": expire,
    }
    return jwt.encode(
        payload,
        settings.jwt_secret,
        algorithm=settings.jwt_algorithm,
    )


def decode_token(token: str) -> dict | None:
    try:
        return jwt.decode(
            token,
            settings.jwt_secret,
            algorithms=[settings.jwt_algorithm],
        )
    except jwt.PyJWTError:
        return None
