API/pyproject.toml
[project]
name = "mentorix-api"
version = "0.1.0"
description = "Mentorix MVP backend"
readme = "API_README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "sqlalchemy>=2.0.30",
  "asyncpg>=0.29.0",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.2",
  "redis>=5.0.7",
  "pgvector>=0.3.2",
  "pymongo>=4.10.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "pypdf>=5.0.0",
  "passlib[bcrypt]>=1.7.4",
  "pyjwt>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
  "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "tests*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
